+++ src/blob/middlewares/blobStorageContext.middleware.ts
@@ -12,7 +12,7 @@ import {
   ValidAPIVersions,
   VERSION
 } from "../utils/constants";
-import { checkApiVersion } from "../utils/utils";
+import { checkApiVersion, validateContainerName } from "../utils/utils";
 
 export default function createStorageBlobContextMiddleware(
   skipApiVersionCheck?: boolean
@@ -102,6 +102,11 @@ export function blobStorageContextMiddleware(
     );
 
     return next(handlerError);
+  }
+
+  // validate conatainer name
+  if (container !== undefined) {
+    validateContainerName(requestID, container);
   }
 
   logger.info(